pyproject.toml
[project]
name = "nonebot-plugin-exchangerate"
version = "0.0.0"
description = "NoneBot2 plugin for currency converter"
authors = [
    {name = "Akirami", email = "Akiramiaya@outlook.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.0rc2", "nonebot-adapter-onebot>=2.1.5", "httpx>=0.23.0"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/A-kirami/nonebot-plugin-exchangerate"
Repository = "https://github.com/A-kirami/nonebot-plugin-exchangerate"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
